--- week4/Calculator/Calculator/Colors.xcassets/darkGray.colorset/Contents.json
@@ -1,23 +1,0 @@
-{
-  "colors" : [
-    {
-      "color" : {
-        "color-space" : "srgb",
-        "components" : {
-          "alpha" : "1.000",
-          "blue" : "0x51",
-          "green" : "0x51",
-          "red" : "0x51"
-        }
-      },
-      "idiom" : "universal"
-    }
-  ],
-  "info" : {
-    "author" : "xcode",
-    "version" : 1
-  },
-  "properties" : {
-    "localizable" : true
-  }
-}
