--- week4/Calculator/Calculator/Colors.xcassets/lightOrange.colorset/Contents.json
@@ -1,23 +1,0 @@
-{
-  "colors" : [
-    {
-      "color" : {
-        "color-space" : "srgb",
-        "components" : {
-          "alpha" : "1.000",
-          "blue" : "10",
-          "green" : "160",
-          "red" : "254"
-        }
-      },
-      "idiom" : "universal"
-    }
-  ],
-  "info" : {
-    "author" : "xcode",
-    "version" : 1
-  },
-  "properties" : {
-    "localizable" : true
-  }
-}
