--- week4/Calculator/Calculator/Extensions/Color+Extension.swift
@@ -1,14 +1,0 @@
-//
-//  Color+Extension.swift
-//  Calculator
-//
-//  Created by Kyungsoo Lee on 2022/08/16.
-//
-
-import SwiftUI
-
-extension Color {
-    static let darkGray = Color("darkGray")
-    static let lightGray = Color("lightGray")
-    static let lightOrange = Color("lightOrange")
-}
