+++ week4/README.md
@@ -1,17 +1,15 @@
-### Week4
-//AC하고 toggleSign할 때 옵셔널때문에 에러. (displayText 맨 앞에 "-" 부호를 바꿔서 stack에 넣는 방식으로 다시 개선해볼 것.
-//% 구현 과정에서 예외처리 확실하게 할 것(사실 inputNumber를 스택에 바로바로 넣었다 뺐다 하면 해결)
-// 전체적인 구현 로직 다시 한 번 생각해볼 것
--> 현재는 연산자가 들어오면 숫자를 stack에 넣는 방식이지만, 연산자가 들어오기 전에 스택에 숫자를 바로바로 갱신해주는 과정으로 코드를 짜면 보다 전체적인 코드가 간결해질 것 같음.(stack을 적극적으로 활용하자.)
+### Week4 & Week5
 
-언래핑 강제로 하는거 줄이기.
-스택의 기능 분리(calculator에서 분리하고, stack타입으로 해서 구조체로 해결하기. ))
-스택을 스택 타입으로Stack ; 
-[String]
-Stack ; [String]
+1. 네이밍 setDigit, setOperation
+2. 로직적인 부분에서 중복 없애기 (displayValue, newValue)
+3. nsNumber로 값을 Decimal로 받을 수 있도록 하기.(,)으로   -> extension으로 해결 ------------수정중------------
+    -> 네이밍 수정
+4. 핸드폰 계산기로 할 때, 어느 일정 개수(9개)까지만 입력되게끔 고치기.
+5. exponential Notation 구현(999,999,999가 넘어가면 지수 표기법으로 표기)----------해결-------------
+    -----------------------수정중--------------
+    -> 네이밍 수정
+6. viewModel에서 buttonLayout 중복 처리 ---------------------해결----------------------------------
 
++
 
-
-연산용 모델을 따로 분리해서 처리하기.
-
-차라리 let symbol을 스태틱처리 해보기.(타입에 고정된 것이라서)
+7. 버튼의 model, viewmodel 어디에 구현할지. 또, 어떤식으로 구현해야 나은 mvvm모델이 될지 고민
