+++ week5/README.md
@@ -1,15 +1,1 @@
 ### Week5
-
-1. 네이밍 setDigit, setOperation
-2. 로직적인 부분에서 중복 없애기 (displayValue, newValue)
-3. nsNumber로 값을 Decimal로 받을 수 있도록 하기.(,)으로   -> extension으로 해결 ------------수정중------------
-    -> 네이밍 수정
-4. 핸드폰 계산기로 할 때, 어느 일정 개수(9개)까지만 입력되게끔 고치기.
-5. exponential Notation 구현(999,999,999가 넘어가면 지수 표기법으로 표기)----------해결-------------
-    -----------------------수정중--------------
-    -> 네이밍 수정
-6. viewModel에서 buttonLayout 중복 처리 ---------------------해결----------------------------------
-
-+
-
-7. 버튼의 model, viewmodel 어디에 구현할지. 또, 어떤식으로 구현해야 나은 mvvm모델이 될지 고민
